drivers/libhid.{c,h}: fix max_report_size from int to size_t

+++ drivers/libhid.c
@@ -57,7 +57,7 @@ static int8_t get_unit_expo(const HIDData_t *hiddata);
 static double exponent(double a, int8_t b);
 
 /* Tweak flag for APC Back-UPS */
-int max_report_size = 0;
+size_t max_report_size = 0;
 
 /* Tweaks for Powercom, at least */
 int interrupt_only = 0;
@@ -159,7 +159,7 @@ static int refresh_report_buffer(reportbuf_t *rbuf, hid_dev_handle_t udev, HIDDa
 	}
 
 	r = comm_driver->get_report(udev, id, rbuf->data[id],
-		max_report_size ? (int)sizeof(rbuf->data[id]):rbuf->len[id]);
+		max_report_size ? sizeof(rbuf->data[id]) : rbuf->len[id]);
 
 	if (r <= 0) {
 		return -1;
